Refactor converter test class (TestLinearInequalityToPenalty)
